2008-06-23  Neal H. Walfield  <neal@gnu.org>
	* object.h (struct object_desc): Change AGE from a uint16_t to a
	uint8_t.
	* ager.c (ager_loop): Set TIMEOUT to a quarter of a second.

+++ viengoos/object.h
@@ -153,7 +153,7 @@ struct object_desc
   struct object_policy policy;
 
   /* The object's age.  */
-  uint16_t age;
+  uint8_t age;
 
   /* The activity to which the *memory* (not the storage) is
      attributed.  If none, then NULL.  (The owner of the storage can
